CMake: Improve iOS bundle identifier handling
Developers could use two different CMake properties / variables to
specify the bundle identifier of an application. One would be written
into the Info.plist file and the other into the Xcode project as a
project property.
The previous implementation forced developers to set both of them to
the same value to have consistency in the project. This is not nice
behavior.

The new logic allows setting only one of the two, which will then be
reused for the one left unspecified. This is more user friendly.

Finally the bundle identifier prefix was never escaped, which could
cause issues during code signing if it had invalid characters.
Now we escape the prefix to ensure the whole identifier is valid.

Amends 4d838dae5a821e9e5f013ba1d5a494ece1b5180e

Pick-to: 6.2 6.3
Task-number: QTBUG-95838
Change-Id: I1448bd643893751c281cc0ea41c763d1bd03360e
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>