CMake: Ensure creation of a unique iOS bundle identifier
Apple ties iOS application bundle identifiers to Apple accounts.
When building against the device SDK, the bundle identifier must be
unique, so that Xcode can register the id with the used developer
account and create an appropriate provisioning profile for the used
device.

If it's not unique (was previously used on a machine with a different
developer account), in the best case you get an error explaining that.
In the worst case you get an error about automatic signing not being
enabled (at least on the command line).

When building an Xcode project for the first time on a fresh machine,
the organization bundle identifier prefix might not be set.
Opening a Qt example project on such a machine might result in
generating the same bundle identifier, leading to the error described
above.

Warn project builders that they should either specify a prefix or an
entire unique identifier.

To ensure building still works in such a case, try to hash the team
id if it is present and add it as an infix to the auto-generated
bundle identifier.

Note that simulator builds are not affected by id uniqueness because
apps deployed to the simulator do not require a provisioning profile
from Apple.

Amends 4d838dae5a821e9e5f013ba1d5a494ece1b5180e

[ChangeLog][iOS][CMake] The build system tries to create a unique
bundle identifier based on the team id if no organization prefix can
be retrieved from Xcode preferences.

Task-number: QTBUG-95838
Change-Id: I6ad10076751acf43064330a6e53d6c4e973ec362
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit a2255a3ab224099eaaf7c3ab37cd6e9c18aa5105)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>