xcb: correctly update the state of m_singlescreen
If the patch is not applied (404c5dc49fca9fdded44df6e7810d3ac8cdd970e),
after the two screens are spliced into a virtual large screen,
unplug the non-primary screen, and then plug it again, the resolution
 size of qscreen update is incorrect (only the size of the main screen).
Because the patch 404c5dc4* is not applied, when the non-primary screen is
unplugged, the interface findScreenForMonitorInfo will return true,
thereby updating the screen information through updateScreen_monitor.
After unplugging the non-main screen, the number of crtc obtained by
the setmonitor interface is 1, and m_singlescreen is true. After
plugging in the non-main screen again, continue to update the screen
information through updateScreen_monitor, but in the setmonitor interface,
since m_singlescreen saves the last value (true), the virtual geometry
information (m_geometry) is not obtained correctly.

Although after patch 404c5dc4* is applied, the screen information class
will be recreated after unplugging and plugging the screen, and the
updateScreen_monitor interface will not be used, but in order to improve
the completeness of the existing interface updateScreen_monitor,
m_singlescreen needs to be correctly updated.

Signed-off-by: wujiang <wujiang@kylinos.cn>