xcb: keep the primary screen at the head of the screen queue
If patch 404c5dc49fca9fdded44df6e7810d3ac8cdd970e is not applied,
when three screens (A, B, C) are in extended mode, where A is the
primary screen, set A and B to be spliced into a virtual screen,
and C remains in extended mode, then the main screen will become C.

In the above experiment, before splicing is set, the three screens
A, B, and C will create screen information in turn. The virtual
screen after A and B splicing will call createScreen_monitor to
create screen information, and screen C will call the
updateScreen_monitor interface to update the screen information.
Because the information of the virtual screen (AB) formed by
splicing A and B will not return to the primary screen, A and B
will be deleted, and AB will be added, so the screen queue (m_screens)
 will change from A->B->C It is C->AB, so C becomes the main screen.

Although the updateScreen_monitor interface will not be used in the
above experiments after patch 404c5dc4 is applied, in order to improve
the completeness of the updateScreen_monitor interface, the main screen
should be kept at the head of the screen queue in all cases.

Signed-off-by: wujiang <wujiang@kylinos.cn>