Update qt_attribution for harfbuzz
Amends d88da0b2b0a8abaaa521bef864378fe1dbf86709 where I forgot
updating the attribution.

Note that the original change was never picked to 6.2 after all,
since that is on an even older version of Harfbuzz, therefore this
one is only picked to 6.3. Given the phase 6.2 is in, I decided it
was better to leave it, since it didn't get the 3.0.0 update either.

Pick-to: 6.3
Task-number: QTBUG-103603
Change-Id: Ie323d931b49158776c2b9fcc4835b69ca4e2f905
Reviewed-by: Lars Knoll <lars.knoll@qt.io>